Move unsaved changes prompt to `ApplicationForm`
